requirements.txt
pandas
requests
gspread
datetime
lxml
Jinja2
python-dotenv
